requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
alembic==1.13.1
redis==4.6.0
fastapi-cache2==0.2.1
pydantic==2.7.4
pydantic-settings==2.3.3
pymysql==1.1.1
SQLAlchemy~=2.0.31
aiomysql==0.2.0
cryptography==42.0.8
websockets~=12.0
requests~=2.32.3